+++ vol1/java-examples/src/main/java/com/heatonresearch/aifh/randomize/LinearCongruentialRandom.java
@@ -185,7 +185,7 @@ public class LinearCongruentialRandom extends AbstractBoxMuller {
      * {@inheritDoc}
      */
     @Override
-    public double nextFloat() {
+    public float nextFloat() {
         return (float) nextDouble();
     }
