+++ vol1/java-examples/src/main/java/com/heatonresearch/aifh/randomize/SecureGenerateRandom.java
@@ -79,7 +79,7 @@ public class SecureGenerateRandom extends AbstractGenerateRandom {
      * {@inheritDoc}
      */
     @Override
-    public double nextFloat() {
+    public float nextFloat() {
         return this.rand.nextFloat();
     }
